[rom] add load access fault exception handler
This updates the ROM exception handler to handle load access faults that
happen as a result of a flash ECC error. When this type of exception is
generated, the exception handler overwrites function argument registers
(a0 and a1) with all 1s to ensure a legal yet invalid value is returned
from the read, since flash erases to all 1s.

Signed-off-by: Tim Trippel <ttrippel@google.com>